Create tests to check if seaport is allowed transfer vouchers
