+++ contracts/protocol/clients/voucher/BosonVoucher.sol
@@ -579,9 +579,8 @@ contract BosonVoucherBase is IBosonVoucher, BeaconClientBase, OwnableUpgradeable
     function callExternalContract(address _to, bytes memory _data) external onlyOwner {
         require(_to != address(0), INVALID_ADDRESS);
 
-        (bool success, bytes memory result) = _to.call(_data);
+        (bool success, ) = _to.call(_data);
 
-        // Handle result
         if (!success) {
             // Reverts with default message
             revert(EXTERNAL_CALL_FAILED);
