+++ test/integration/seaport/fixtures.js
@@ -1,43 +1,11 @@
 const hre = require("hardhat");
 const ethers = hre.ethers;
-const { constants, BigNumber, utils } = ethers;
+const { constants, BigNumber } = ethers;
 const { abi } = require("./artifacts/contracts/Seaport.sol/Seaport.json");
 const { getOfferOrConsiderationItem, calculateOrderHash } = require("./utils");
-const { expect, assert } = require("chai");
-const { keccak256 } = require("ethers/lib/utils");
+const { expect } = require("chai");
 
 const SEAPORT_ADDRESS = "0x00000000000001ad428e4906aE43D8F9852d0dD6"; // 1.4
-
-const orderType = {
-  OrderComponents: [
-    { name: "offerer", type: "address" },
-    { name: "zone", type: "address" },
-    { name: "offer", type: "OfferItem[]" },
-    { name: "consideration", type: "ConsiderationItem[]" },
-    { name: "orderType", type: "uint8" },
-    { name: "startTime", type: "uint256" },
-    { name: "endTime", type: "uint256" },
-    { name: "zoneHash", type: "bytes32" },
-    { name: "salt", type: "uint256" },
-    { name: "conduitKey", type: "bytes32" },
-    { name: "counter", type: "uint256" },
-  ],
-  OfferItem: [
-    { name: "itemType", type: "uint8" },
-    { name: "token", type: "address" },
-    { name: "identifierOrCriteria", type: "uint256" },
-    { name: "startAmount", type: "uint256" },
-    { name: "endAmount", type: "uint256" },
-  ],
-  ConsiderationItem: [
-    { name: "itemType", type: "uint8" },
-    { name: "token", type: "address" },
-    { name: "identifierOrCriteria", type: "uint256" },
-    { name: "startAmount", type: "uint256" },
-    { name: "endAmount", type: "uint256" },
-    { name: "recipient", type: "address" },
-  ],
-};
 
 const seaportFixtures = async () => {
   const seaport = await ethers.getContractAt(abi, SEAPORT_ADDRESS);
